Refactoring code.
